==> Scripts de Tratamento/7 - carrega_arquivos_no_banco.ipynb <==
import os
import pandas as pd
from sqlalchemy import create_engine, text

engine = create_engine("postgresql+psycopg2://postgres:1234@localhost:5433/Dados_TCC")

folder_path = r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC\Scripts\Arquivos Finais"


files = [f for f in os.listdir(folder_path) if f.endswith(('.xlsx', '.xls'))]

for file in files:
    file_path = os.path.join(folder_path, file)
    
    df = pd.read_excel(file_path)
    
    table_name = os.path.splitext(file)[0]
    
    df.to_sql(table_name, engine, if_exists='replace', index=False)
