==> Scripts de Tratamento/8 - extração_para_bi.ipynb <==
import os
import pandas as pd
from sqlalchemy import create_engine, inspect

output_folder = r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC\Scripts\Arquivos para BI"
os.makedirs(output_folder, exist_ok=True)  

engine = create_engine("postgresql+psycopg2://postgres:1234@localhost:5433/Dados_TCC")

inspector = inspect(engine)
tables = inspector.get_table_names()

for table_name in tables:
    df = pd.read_sql_table(table_name, engine)
    
    file_path = os.path.join(output_folder, f"{table_name}.xlsx")
    
    df.to_excel(file_path, index=False)
    
    print(f"Tabela '{table_name}' exportada para '{file_path}' ")
