==> Scripts de Tratamento/9 - indicador final.ipynb <==
import pandas as pd
from datetime import datetime

# Estações

cimehgo = pd.read_excel(r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC 2\Scripts\Arquivos Finais\CIMEHGO_FINAL.xlsx")
inmet = pd.read_excel(r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC 2\Scripts\Arquivos Finais\INMET_FINAL.xlsx")

cimehgo_p = cimehgo.copy()
inmet_p = inmet.copy()

cimehgo_p = cimehgo_p.rename(columns={
    "ESTAÇÃO": "Estacao",
    "DATA/HORA(LOCAL)": "Data",
    "CHUVA(MM)": "Chuva_mm"
})

inmet_p = inmet_p.rename(columns={
    "ESTACAO": "Estacao",
    "Data": "Data",
    "Chuva(mm)": "Chuva_mm",                   
})
inmet_p["Chuva_mm"]   = pd.to_numeric(inmet_p["Chuva_mm"], errors="coerce").fillna(0)
cimehgo_p["Chuva_mm"] = pd.to_numeric(cimehgo_p["Chuva_mm"], errors="coerce").fillna(0)

inmet_p["Hora"] = pd.to_datetime(inmet_p["Hora"], format="%H:%M", errors="coerce").dt.time
inmet_p["Data"] = pd.to_datetime(inmet_p["Data"], errors="coerce")
inmet_p["Data"] = inmet_p.apply(lambda row: pd.Timestamp.combine(row["Data"], row["Hora"]), axis=1)

cimehgo_p = cimehgo_p[["Estacao", "Data", "Chuva_mm", "LATITUDE", "LONGITUDE"]]
inmet_p = inmet_p[["Estacao", "Data", "Chuva_mm", "LATITUDE", "LONGITUDE"]]

final_est = pd.concat([cimehgo_p, inmet_p], ignore_index=True)

final_est.to_csv(r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC 2\Scripts\Arquivos para BI\DADOS_UNIFICADOS_ESTACOES.csv", index=False, encoding="latin1")

# Ocorrências

lp = pd.read_excel(r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC 2\Scripts\Arquivos Finais\Locais_por_Data_GEOCODED_FINAL.xlsx")
rp = pd.read_excel(r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC 2\Scripts\Arquivos Finais\RELATOS_GEOCODED_FINAL.xlsx")
bp = pd.read_excel(r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC 2\Scripts\Arquivos Finais\BOMBEIROS_GEOCODED_FINAL.xlsx")

locais = lp.copy()
relatos = rp.copy()
bombeiros = bp.copy()
bombeiros = bombeiros.rename(columns={"ENDEREÇO": "LOCAL"})

locais = locais[["DATA", "LOCAL", "Estação atribuída", "LATITUDE", "LONGITUDE"]]
relatos = relatos[["DATA", "LOCAL", "Estação atribuída", "LATITUDE", "LONGITUDE"]]
bombeiros = bombeiros[["DATA", "LOCAL", "Estação atribuída", "LATITUDE", "LONGITUDE"]]

locais["Fonte"] = "Noticias"
relatos["Fonte"] = "Relatos"
bombeiros["Fonte"] = "Bombeiros"

def parse_date(x):
    if pd.isna(x):
        return pd.NaT
    
    s = str(x).strip()

    # Remover dia da semana (ex.: "Sunday, January 7, 2024")
    if "," in s:
        partes = s.split(",", 1)  # remove só a 1ª vírgula
        if len(partes) == 2:
            s = partes[1].strip()  # vira "January 7, 2024"

    # Tentativa 1: conversão automática do pandas
    try:
        return pd.to_datetime(s, errors="raise", infer_datetime_format=True)
    except:
        pass

    # Tentativa 2: formato textual American ("April 7 2024" / "April 7, 2024")
    try:
        return datetime.strptime(s, "%B %d %Y")
    except:
        pass

    try:
        return datetime.strptime(s, "%B %d, %Y")
    except:
        pass

    # Tentativa final: retornar NaT
    return pd.NaT

locais["DATA"] = locais["DATA"].apply(parse_date)
relatos["DATA"] = relatos["DATA"].apply(parse_date)
bombeiros["DATA"] = bombeiros["DATA"].apply(parse_date)

final_ocor = pd.concat([locais, relatos, bombeiros], ignore_index=True)

final_ocor

final_ocor.to_excel(r"C:\Users\thiag\OneDrive\Área de Trabalho\TCC 2\Scripts\Arquivos para BI\DADOS_GEOCODED_UNIFICADOS.xlsx", index=False)

# Análise do Indicador 

# Para cada linha de final_ocor, veja se tem uma Estação atribuída que está dentro da lista de Estacao de final_est
valid_stations = final_est["Estacao"].unique()
final_ocor_filt = final_ocor[ final_ocor["Estação atribuída"].isin(valid_stations) ]

final_ocor_filt = final_ocor_filt.drop_duplicates()

df_alag = final_ocor.copy()
df_chuva = final_est.copy()

df_chuva['Data'] = pd.to_datetime(df_chuva['Data'])
df_chuva['data_dia'] = df_chuva['Data'].dt.date

df_alag['DATA'] = pd.to_datetime(df_alag['DATA'])
df_alag['data_dia'] = df_alag['DATA'].dt.date

df_alag = df_alag.rename(columns={'Estação atribuída': 'Estacao'})

estacoes_validas = set(df_chuva['Estacao'].unique())
df_alag_filtrado = df_alag[df_alag['Estacao'].isin(estacoes_validas)]

df_expandido = pd.merge(
    df_alag_filtrado,
    df_chuva,
    on=['Estacao', 'data_dia'],
    how='left'
)

estacoes_ausentes = (
    set(df_alag['Estacao'].unique())
    - set(df_chuva['Estacao'].unique())
)

# Tem no df_alag, mas nao no df_chuva
estacoes_ausentes

df_expandido


estatisticas_eventos = (
    df_expandido
    .groupby(['LOCAL', 'DATA', 'Estacao'])['Chuva_mm']
    .agg(['min','max','mean'])
    .reset_index()
)

estatisticas_eventos

df_chuva['data'] = pd.to_datetime(df_chuva['Data'])
df_alag['data'] = pd.to_datetime(df_alag['DATA'])

df_chuva['Data'] = pd.to_datetime(df_chuva['Data'])
df_alag['DATA'] = pd.to_datetime(df_alag['DATA'])

df_alag_ren = df_alag.rename(columns={
    'Estação atribuída': 'Estacao',
    'DATA': 'Data'
})

#df_merge

"""df_merge = pd.merge(
    df_chuva,
    df_alag_ren[['Data', 'Estacao']], 
    on=['Estacao', 'Data'],
    how='inner'
)"""

data_u = final_ocor_filt

# Para cada um desses pares de data e estação, ver no df final_est todas as ocorrencias com mesma data (vai precisar normalizar a hora) para pegar os valores de chuva (mm) do dia

# Fazer três colunas, máxima, minima e média de chuva (mm) capturadas pelos pluviometros naquela data quando foi registrado um alagamento, com isso a gente pode estimar que
# caso tenha mais de uma ocorrência no mesmo local, os valores observados podem ser considerados indicadores de alagamento para a região da estação.

# No TCC isso precisa ser muito bem explicado para ficar obvio que os dados atuais são apenas estimativas e que um indicador de verdade precisa de muitos mais dados para ser formado
